refactor: remove `RecipientAccountSelector`
