+++ packages/shared/components/modals/index.js
@@ -5,4 +5,3 @@ export { default as Modal } from './Modal.svelte'
 export { default as NodeActionsMenu } from './NodeActionsMenu.svelte'
 export { default as ProfileActionsModal } from './ProfileActions.svelte'
 export { default as ProposalsDetailsMenu } from './ProposalsDetailsMenu.svelte'
-export { default as RecipientAccountSelector } from './RecipientAccountSelector.svelte'
